Refactor java.rs into separate files
* Instead of 1 600 line file, a few more cohesive 100 line files
