+++ src/lang/java/class_def.rs
@@ -1,0 +1,163 @@
+use crate::java::method_def::parse_method;
+use crate::java::modifier::{find_modifier, parse_modifiers, Modifier};
+use crate::java::util::fold_vec;
+use crate::java::vartype::find_type;
+use crate::parse::AST;
+use crate::prophet::*;
+
+/// Handles class definition portions of the Java language, like fields,
+/// orchestrating the class body, etc.
+
+/// Parse a single class/interface/annotation/what have you's AST
+pub(crate) fn parse_class(
+    ast: &AST,
+    package: &str,
+    path: &str,
+) -> Option<ClassOrInterfaceComponent> {
+    // Get container info
+    let instance_type = match ast.find_child_by_type(&["class", "interface", "enum", "annotation"])
+    {
+        Some(r#type) => match &*r#type.value {
+            "interface" => InstanceType::InterfaceComponent,
+            _ => InstanceType::ClassComponent,
+        },
+        None => InstanceType::ClassComponent,
+    };
+    let instance_name = match ast.find_child_by_type(&["identifier"]) {
+        Some(identifier) => identifier.value.clone(),
+        None => "".into(),
+    };
+
+    let component = ComponentInfo {
+        path: path.into(),
+        package_name: package.into(),
+        instance_name: format!(
+            "{}::{}",
+            instance_name,
+            match instance_type {
+                InstanceType::InterfaceComponent => "InterfaceComponent",
+                _ => "ClassComponent",
+            }
+        ),
+        instance_type: instance_type,
+    };
+
+    // Define default values
+    let mut stereotype = ContainerStereotype::Entity;
+    let mut fields = vec![];
+    let mut constructors = vec![];
+    let mut methods = vec![];
+    let mut modifier = Modifier::new();
+
+    // Find bounds
+    let start;
+    let end;
+    if let Some(span) = ast.span {
+        start = span.0 as i32;
+        end = span.1 as i32;
+    } else {
+        start = 0;
+        end = 0;
+    }
+
+    // Generate the implementation data
+    for member in ast.children.iter() {
+        match &*member.r#type {
+            "modifiers" => {
+                modifier = parse_modifiers(member, &*component.path, &*component.package_name)
+            }
+            "class_body" | "interface_body" | "enum_body" | "annotation_body" => {
+                parse_class_body(
+                    member,
+                    &component,
+                    &mut constructors,
+                    &mut methods,
+                    &mut fields,
+                );
+            }
+            unknown_type => {
+                eprintln!(
+                    "{} tag unhandled. This may or may not be an issue.",
+                    unknown_type
+                );
+            }
+        };
+    }
+
+    // Assemble the return value from what was extracted from the tree
+    Some(ClassOrInterfaceComponent {
+        component: ContainerComponent {
+            component,
+            accessor: modifier.accessor,
+            stereotype: stereotype,
+            methods,
+            container_name: instance_name,
+            line_count: end - start + 1,
+        },
+        declaration_type: ContainerType::Class,
+        annotations: modifier.annotations,
+        constructors: fold_vec(constructors),
+        field_components: fold_vec(fields),
+    })
+}
+
+/// Parses the members of a class to find members, methods, blocks, etc.
+fn parse_class_body(
+    ast: &AST,
+    component: &ComponentInfo,
+    constructors: &mut Vec<MethodComponent>,
+    methods: &mut Vec<MethodComponent>,
+    fields: &mut Vec<FieldComponent>,
+) {
+    // Traverse body
+    for member in ast.children.iter() {
+        match &*member.r#type {
+            "constructor_declaration" | "static_initializer" => {
+                constructors.push(parse_method(member, component))
+            }
+            "method_declaration" => methods.push(parse_method(member, component)),
+            "field_declaration" => fields.push(parse_field(member, component)),
+            "class_declaration"
+            | "interface_declaration"
+            | "enum_declaration"
+            | "annotation_declaration" => { /* None, since these were extracted + handled elsewhere */
+            }
+            unknown => eprintln!(
+                "Attempting to parse {}, not currently supported. TODO implement fully!",
+                unknown
+            ),
+        }
+    }
+}
+
+/// Parses a single field in a class
+fn parse_field(ast: &AST, component: &ComponentInfo) -> FieldComponent {
+    let variables = ast
+        .find_all_children_by_type(&["variable_declarator"])
+        .get_or_insert(vec![])
+        .iter()
+        .map(|var_decl| var_decl.find_child_by_type(&["identifier"]))
+        .filter(|var_ident| var_ident.is_some())
+        .map(|identifier| identifier.unwrap().value.clone())
+        .collect();
+    let modifier = find_modifier(ast, &*component.path, &*component.package_name);
+    let r#type = find_type(ast);
+
+    // TODO: How to handle field_name, default_value?
+    FieldComponent {
+        component: ComponentInfo {
+            path: component.path.clone(),
+            package_name: component.package_name.clone(),
+            instance_name: component.instance_name.clone(),
+            instance_type: InstanceType::FieldComponent,
+        },
+        annotations: modifier.annotations,
+        variables,
+        field_name: String::new(),
+        accessor: modifier.accessor,
+        is_static: modifier.is_static,
+        is_final: modifier.is_final,
+        default_value: String::new(),
+        r#type,
+    }
+}
